♻️ refactor(task/schedule): 移除自定义ID生成器，改用xorm.Model自动生成主键
- Task/Schedule模型移除model.Base和BeforeInsert，改用xorm.Model
- Task.Stop改为*time.Time，支持空值判断超时中断任务
- Maximum/Times字段由uint32改为uint64
- 列常量由gox.Column改为普通字符串，移除相关依赖
- Database.DB重命名为Engine，统一字段命名
- 升级goexl/task依赖至v0.3.0，迁移版本号升级至3

+++ internal/internal/internal/kernel/task.go
@@ -36,11 +36,11 @@ func (t *Task) Subtype() task.Type {
 	return t.task.Subtype
 }
 
-func (t *Task) Maximum() uint32 {
+func (t *Task) Maximum() uint64 {
 	return t.task.Maximum
 }
 
-func (t *Task) Times() uint32 {
+func (t *Task) Times() uint64 {
 	return t.task.Times
 }
